backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# AWS Services
boto3==1.34.34
botocore==1.34.34

# Vector Store
chromadb==0.4.22
sentence-transformers==2.3.1

# AI/ML
anthropic==0.18.1
langchain==0.1.6
langchain-community==0.0.19

# Data Processing
pandas==2.2.0
numpy==1.26.3
prophet==1.1.5

# Report Generation
weasyprint==60.2
jinja2==3.1.3

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.3.0
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
hypothesis==6.98.3
